Addind stuff related to buttons

+++ lib/button_manager/button_manager.h
@@ -5,15 +5,14 @@
 
 class ButtonManager {
 public:
-    ButtonManager(int leftButtonPin, int rightButtonPin, int playPauseButtonPin);
+    ButtonManager(unsigned short leftButtonPin, unsigned short rightButtonPin, unsigned short playPauseButtonPin);
     void begin();
     unsigned short checkButtons();
 
 private:
-    int leftButtonPin;
-    int rightButtonPin;
-    int playPauseButtonPin;
-
+    unsigned short leftButtonPin;
+    unsigned short rightButtonPin;
+    unsigned short playPauseButtonPin;
 };
 
 #endif //BUTTON_MANAGER_HEADER
